+++ src/Wahren.AbstractSyntaxTree/Parser/Solution/IStringResolver.cs
@@ -3,7 +3,7 @@
 public interface IStringResolver<T>
     where T : IIdEnumerator
 {
-    void Register(nuint resultId, uint id, ReadOnlySpan<char> key);
+    void Register(uint resultId, uint id, ReadOnlySpan<char> key);
     ref Result TryGet(ReadOnlySpan<char> key, out uint value);
     bool Exists(ReadOnlySpan<char> key);
